+++ Project 1/header/bptree.h
@@ -35,6 +35,8 @@ Header for nodes -> 11B
 #define MAX_LEAF_KEYS     (((NODE_SIZE - NODE_HDR_SIZE) / (KEY_SIZE + RECORD_POINTER_SIZE)) - 1)
 #define MIN_INTERNAL_KEYS ((MAX_INTERNAL_KEYS) / 2)
 #define MIN_LEAF_KEYS     ((MAX_LEAF_KEYS + 1) / 2)
+#define MAX_INT_CHILDREN (MAX_INTERNAL_KEYS + 1)
+
 
 typedef struct  {
     uint16_t key_count;
